rename debouncedInput to debouncedValue, merge getPopularSearchTerms and gePopu...Options.ts in a single file. Change the getPopularSearchTerms mock in the test.
